Solved issue with tests (now they run sequentially) due to port usage. Reorganized tests and eliminated duplicates.
