+++ llmservice/__mocks__/@google/generative-ai.js
@@ -1,0 +1,21 @@
+function GoogleGenerativeAI() {
+    return {
+        getGenerativeModel: () => ({
+        generateContent: () => ({
+            response: {
+            candidates: [
+                {
+                content: {
+                    parts: [{ text: "llmAnswer" }],
+                },
+                },
+            ],
+            },
+        }),
+        }),
+    };
+    }
+
+module.exports = {
+  GoogleGenerativeAI,
+};
